services/api/requirements.txt
fastapi
uvicorn[standard]
boto3
pydantic
pydantic-settings
psycopg2-binary
types-psycopg2
httpx
pytest
